Refactor API layer in front-end
The access to the Rust backend as well providing a store to cache the
list of chats are no covered in the API layer of the front-end.
Components depend on it to interact with the backend and get the correct
data. Caching is part of that API layer.
